first steps towards blender 4
Import of `.obj` files is now natively supported and much faster. This
is a huge speed improvement for `importMeshes.py` and necessary for our
many movies.

The "Principled BSDF" has received a rewrite, some of the argument names
for the nodes changed, mostly in the `utilsMaterials.py` file.

I only fixed the parts that are necessary for our movies, there might be
other issues hidden in the code.